refactor: Seperate reusable from behavioral parts
Extract parts of the code that can be used in different strategies for writing ltr<->rtl css into their
own `core` submodule so that they can be shared.

Also moves implementation-specific background-image and background-position regexes
from inside the relevant functions into the specific implementation
