+++ src/core/calculateNewTranslate.js
@@ -1,0 +1,5 @@
+import flipSign from './flipSign'
+
+export default function calculateNewTranslate(match, prefix, offset, suffix) {
+  return prefix + flipSign(offset) + suffix
+}
