+++ src/core/includes.js
@@ -1,0 +1,3 @@
+export default function includes(inclusive, inclusee) {
+  return inclusive.indexOf(inclusee) !== -1
+}
