+++ src/core/isNullOrUndefined.js
@@ -1,0 +1,3 @@
+export default function isNullOrUndefined(val) {
+  return val === null || typeof val === 'undefined'
+}
