+++ src/core/isNumber.js
@@ -1,0 +1,3 @@
+export default function isNumber(val) {
+  return typeof val === 'number'
+}
