+++ src/core/isObject.js
@@ -1,0 +1,3 @@
+export default function isObject(val) {
+  return val && typeof val === 'object'
+}
