+++ src/core/isString.js
@@ -1,0 +1,3 @@
+export default function isString(val) {
+  return typeof val === 'string'
+}
